+++ initData/policy.js
@@ -230,4 +230,18 @@ module.exports = [{
     "type": "ADMIN_MENU",
     "path": "/ott/channels/list",
     "method": "get"
+},{
+    "_id": "510015000000000000000001",
+    "group": "專欄管理",
+    "desc": "新增專欄 - 特輯版型",
+    "type": "ADMIN_MENU",
+    "path": "/column/specialchannel/create",
+    "method": "get"
+},{
+    "_id": "510015000000000000000002",
+    "group": "專欄管理",
+    "desc": "專欄列表 - 特輯版型",
+    "type": "ADMIN_MENU",
+    "path": "/column/specialchannel",
+    "method": "get"
 }];