+++ CHANGELOG.md
@@ -5,6 +5,7 @@ NOWnews-api change logs
 ## [Unreleased]
 
 ### Changed
+- 移除api-web端點 /news中的敏感資訊 @appleoxxo
 - 修改admin端點 /indexpage 內carousel 新增feedFrom欄位資料 @appleoxxo
 - 調整 Role 新增修改時可帶 `level`, `SuperiorRoles` 欄位 @esbb48
 - 在發布與送審時，確認更新者的身份 @esbb48
@@ -18,6 +19,8 @@ NOWnews-api change logs
 - 新增中央社新聞匯入 @appleoxxo
 - 加入 imagelab 縮圖網址與在 Image mode 加入 sizeFormat 虛擬欄位 @SimonSun
 - 增加廠商專屬的列表「我的新聞(廠商)」權限@esbb48
+- 加入特輯版型業配專欄 API， initData policy.js 加入 `專欄管理(專欄列表 - 特輯版型)` 與 `專欄管理(新增專欄 - 特輯版型)` @SimonSun
+- 加入從 excel 匯入新聞關鍵字工具，加入 `convert-excel-to-json` node module @SimonSun
 - 增加欄位 `level`, `SuperiorRoles` 至 Role 表 @esbb48
 - 增加端點 `/roles/:id/reviewers` @esbb48